+++ changelog.md
@@ -330,6 +330,9 @@
 
 - Added `jsconsole.dir`, `jsconsole.dirxml`, `jsconsole.timeStamp`.
 
+- Added `std/smartptrs`, use `-d:nimExperimentalSmartptrs` 
+  to enable this experimental module.
+
 - Added dollar `$` and `len` for `jsre.RegExp`.
 
 - Added `std/tasks`.